// frontend/src/api/generated/models/aulaDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SIFES API
 * API do Sistema do Instituto Federal de Esperantina — Diário de Aula
 * OpenAPI spec version: 1.0
 */
import type { AulaDtoEstado } from './aulaDtoEstado';
import type { AulaDtoEstadoOverride } from './aulaDtoEstadoOverride';

export interface AulaDto {
  id: string;
  materiaId: string;
  data: string;
  horaInicio: string;
  horaFim: string;
  /** @nullable */
  titulo: string | null;
  /** @nullable */
  descricao: string | null;
  /** @nullable */
  estadoOverride: AulaDtoEstadoOverride;
  estado: AulaDtoEstado;
}